UPSTREAM: mt76: add utility functions for deferring work to a kernel thread
In order to avoid keeping work like tx scheduling pinned to the CPU it was
scheduled from, it makes sense to switch from tasklets to kernel threads.

Unlike a workqueue, this one only allows one fixed worker function to be
executed by the worker thread. Because of that, there is less locking
and less code for scheduling involved.
This is important because the tx worker is scheduled often in a hot path

BUG=b:178754244
TEST=build and boot on asurada

(cherry picked from commit 30bdd69246ea9726a7688ef02c60b1ec0ae66d38)
Signed-off-by: Sean Wang <objelf@gmail.com>
Signed-off-by: Felix Fietkau <nbd@nbd.name>
Change-Id: I3dd44d32a0d81649a551cd0712fd60cc0b1e36e9