UPSTREAM: mt76: mt7615: significantly reduce interrupt load
On 7615 and newer, DMA completion only triggers unmap, but not free of queued
skbs, since pointers to packets are queued internally.
Because of that, there is no need to process the main data queue immediately
on DMA completion.
To improve performance, mask out the DMA data queue completion interrupt and
process the queue only when we receive a txfree event.
This brings the number of interrupts under load down to a small fraction.

BUG=b:178754244
TEST=build and boot on asurada

(cherry picked from commit e17d794587298924ba1aed7c642492e170f39d85)
Signed-off-by: Sean Wang <objelf@gmail.com>
Signed-off-by: Felix Fietkau <nbd@nbd.name>
Change-Id: Icdaf631f8ada68a1b5c77a6f3e590128f1b85145