UPSTREAM: mt76: mt7915: fix processing txfree events
In the MT7915 info, the fields for the WLAN index / queue overlap with the
token id, and the MT_TX_FREE_PAIR bit indicates, which one is present.

If MT_TX_FREE_PAIR is set, skip processing the token index, since the data
will not be valid.

This fixes accidentally freeing tokens which are still in use by the
hardware with a wrong station pointer.

BUG=b:178754244
TEST=build and boot on asurada

(cherry picked from commit 6425791d350301b50f5e0332007723de7710d07e)
Signed-off-by: Sean Wang <objelf@gmail.com>
Signed-off-by: Felix Fietkau <nbd@nbd.name>
Change-Id: Ib925251eb32748525b0674ad13793fd2e7551346